+++ spec/integration/braintree/transaction_spec.rb
@@ -3642,6 +3642,182 @@ describe Braintree::Transaction do
         result.errors.for(:transaction).on(:ships_from_postal_code)[0].code.should == Braintree::ErrorCodes::Transaction::ShipsFromPostalCodeInvalidCharacters
       end
     end
+
+    context "network_transaction_id" do
+      it "receives network_transaction_id for visa transaction" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :credit_card => {
+            :number => Braintree::Test::CreditCardNumbers::Visa,
+            :expiration_date => "05/2009"
+          },
+          :amount => "10.00",
+        )
+        result.success?.should == true
+        result.transaction.network_transaction_id.should_not be_nil
+      end
+    end
+
+    context "external vault" do
+      it "returns a validation error if used with an unsupported instrument type" do
+        customer = Braintree::Customer.create!
+        result = Braintree::PaymentMethod.create(
+          :payment_method_nonce => Braintree::Test::Nonce::PayPalFuturePayment,
+          :customer_id => customer.id
+        )
+        payment_method_token = result.payment_method.token
+
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :customer_id => customer.id,
+          :payment_method_token => payment_method_token,
+          :external_vault => {
+            :status => Braintree::Transaction::ExternalVault::Status::WillVault,
+          },
+          :amount => "10.00",
+        )
+        result.success?.should == false
+        result.errors.for(:transaction)[0].code.should == Braintree::ErrorCodes::Transaction::PaymentInstrumentWithExternalVaultIsInvalid
+      end
+
+      it "reject invalid status" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :credit_card => {
+            :number => Braintree::Test::CreditCardNumbers::MasterCard,
+            :expiration_date => "05/2009"
+          },
+          :external_vault => {
+            :status => "not_valid",
+          },
+          :amount => "10.00",
+        )
+        result.success?.should == false
+        result.errors.for(:transaction).for(:external_vault).on(:status)[0].code.should == Braintree::ErrorCodes::Transaction::ExternalVault::StatusIsInvalid
+      end
+
+      context "Visa" do
+        it "accepts status" do
+          result = Braintree::Transaction.create(
+            :type => "sale",
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::Visa,
+              :expiration_date => "05/2009"
+            },
+            :external_vault => {
+              :status => Braintree::Transaction::ExternalVault::Status::WillVault,
+            },
+            :amount => "10.00",
+          )
+          result.success?.should == true
+          result.transaction.network_transaction_id.should_not be_nil
+        end
+
+        it "accepts previous_network_transaction_id" do
+          result = Braintree::Transaction.create(
+            :type => "sale",
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::Visa,
+              :expiration_date => "05/2009"
+            },
+            :external_vault => {
+              :status => Braintree::Transaction::ExternalVault::Status::Vaulted,
+              :previous_network_transaction_id => "123456789012345",
+            },
+            :amount => "10.00",
+          )
+          result.success?.should == true
+          result.transaction.network_transaction_id.should_not be_nil
+        end
+
+        it "rejects non-vaulted status with previous_network_transaction_id" do
+          result = Braintree::Transaction.create(
+            :type => "sale",
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::Visa,
+              :expiration_date => "05/2009"
+            },
+            :external_vault => {
+              :status => Braintree::Transaction::ExternalVault::Status::WillVault,
+              :previous_network_transaction_id => "123456789012345",
+            },
+            :amount => "10.00",
+          )
+          result.success?.should == false
+          result.errors.for(:transaction).for(:external_vault).on(:status)[0].code.should == Braintree::ErrorCodes::Transaction::ExternalVault::StatusWithPreviousNetworkTransactionIdIsInvalid
+        end
+
+        it "rejects invalid previous_network_transaction_id" do
+          result = Braintree::Transaction.create(
+            :type => "sale",
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::Visa,
+              :expiration_date => "05/2009"
+            },
+            :external_vault => {
+              :status => Braintree::Transaction::ExternalVault::Status::Vaulted,
+              :previous_network_transaction_id => "not_and_valid_id",
+            },
+            :amount => "10.00",
+          )
+          result.success?.should == false
+          result.errors.for(:transaction).for(:external_vault).on(:previous_network_transaction_id)[0].code.should == Braintree::ErrorCodes::Transaction::ExternalVault::PreviousNetworkTransactionIdIsInvalid
+        end
+      end
+
+      context "Non-Visa" do
+        it "accepts status" do
+          result = Braintree::Transaction.create(
+            :type => "sale",
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::MasterCard,
+              :expiration_date => "05/2009"
+            },
+            :external_vault => {
+              :status => Braintree::Transaction::ExternalVault::Status::WillVault,
+            },
+            :amount => "10.00",
+          )
+          result.success?.should == true
+          result.transaction.network_transaction_id.should be_nil
+        end
+
+        it "accepts blank previous_network_transaction_id" do
+          result = Braintree::Transaction.create(
+            :type => "sale",
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::MasterCard,
+              :expiration_date => "05/2009"
+            },
+            :external_vault => {
+              :status => Braintree::Transaction::ExternalVault::Status::Vaulted,
+              :previous_network_transaction_id => "",
+            },
+            :amount => "10.00",
+          )
+          result.success?.should == true
+          result.transaction.network_transaction_id.should be_nil
+        end
+
+        it "rejects previous_network_transaction_id" do
+          result = Braintree::Transaction.create(
+            :type => "sale",
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::Discover,
+              :expiration_date => "05/2009"
+            },
+            :external_vault => {
+              :status => Braintree::Transaction::ExternalVault::Status::Vaulted,
+              :previous_network_transaction_id => "123456789012345",
+            },
+            :amount => "10.00",
+          )
+          result.success?.should == false
+          result.errors.for(:transaction).for(:external_vault).on(:previous_network_transaction_id)[0].code.should == Braintree::ErrorCodes::Transaction::ExternalVault::CardTypeIsInvalid
+        end
+      end
+
+    end
   end
 
   describe "self.create!" do