+++ CHANGELOG.md
@@ -3,12 +3,12 @@
 * Fix issue where environment was not automatically converted from a string to a symbol (#148, #164)
 * Fix issue where logger could not be set on gateway instance
 * Fix dispute results in transactions not showing the correct status sometimes
+* Whitelist usage of `external_vault` in transaction create.
+* Visa transactions will now contain a `network_transaction_id` in the response
 
 # 2.90.0
 * Add `processor_response_code` and `processor_response_text` to authorization adjustments subfield in transaction response.
 * Stop restricting `transaction_source` values for Disputes search, allow searching by OAuth Application client_id
-* Add `network_transaction_id` as new field on subfield transaction response.
-* Whitelist usage of `external_vault` in transaction create.
 * Add support for Samsung Pay
 
 # 2.89.0